all: drop support for additionalSecrets
We support the secretMount alternative for 2 releases already.
The migration is easy. Substitute any occurrence of:

```
additionalSecrets:
  volumename:
    secretName: ...
    mountPath: ...
    subPath: ...
```

with:

```
secretMounts:
  - secretName: ...
    mountPath: ...
    subPath: ...
    volumeName: volumename  # volumeName is optional in most cases
```

Renaming additionalSecrets to secretMount is a cosmetic change. The
new name seems to better reflect the function of this section.

Avoiding user-input keys in the yaml object (here: the volume name)
is more consistent with the overall kubernetes best practices.
User input is always passed via yaml values, not via yaml keys.
One of the consequences of this change is improved support for
kubernetes kustomize. Secrets can now be referred via the
fieldSpecs.path fields in kustomize.